Make PathCache stylus resolving a little more like stylus standalone
When running stylus outside of stylus-loader it synchronously resolves
files while evaluating. While it evaluates as it imports a file, it
adds the context (its directory path) of that file to the stylus
instance's `paths` array. `paths` is used to look up relative paths by
iterating from the end of the array to the start.

`stylus-loader` since we added PathCache, left paths alone. This was an
oversight that leads to relative paths not working (unless stylus can't
find it and we fall back to resolving with webpack). `stylus-loader`
can't modify paths like stylus does since it can potentially resolve
multiple dependencies at once. We can concat and create separate
`paths` copies. This change does that by appending the context variable
that for stylus resolving PathCache was otherwise throwing away.

This doesn't accurately follow stylus yet. If you had a tree of stylus
files depending on each other `paths` would have a layer for each
context as it goes deeper.

So this change just represents the last context and not the contexts in
between. This should hopefully cover 99% of relative path use.